% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PractData}
\alias{PractData}
\title{Example Dataset: A simulated dataset for 2SRI-F method}
\format{
A data frame with \code{n} rows and \code{p} variables:
\describe{
\item{time}{Time-to-event variable.}
\item{status}{Censoring indicator: 1 for event, 0 for censored.}
\item{trt}{Treatment indicator: 1 for treated, 0 for control.}
\item{x1}{Covariate 1.}
\item{x2}{Covariate 2.}
}
}
\source{
Simulated data.
}
\usage{
PractData
}
\description{
A brief description of your dataset.
}
\keyword{datasets}
