% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxiv_seq.R
\name{print.coxiv_seq}
\alias{print.coxiv_seq}
\title{Print method for coxiv_seq objects}
\usage{
\method{print}{coxiv_seq}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\description{
Print method for coxiv_seq objects
}
