% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.coxivomom.R
\name{print.coxivomom}
\alias{print.coxivomom}
\title{Print the output from a \code{coxivomom} model fit.}
\usage{
\method{print}{coxivomom}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of the class "\code{coxivomom}", usually created by \code{coxiv_omom()}.}

\item{digits}{the number of significant digits to display for coefficients and hazard ratios.}

\item{...}{additional arguments (currently unused).}
}
\description{
Provide succinct and organized output for a \code{coxivomom} model, such as \code{coxiv_omom()}.
}
\examples{
print(coxiv_omom(Surv(time, censor) ~ x, data = dat))

}
