% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.coxivtsrif.R
\name{print.coxivtsrif}
\alias{print.coxivtsrif}
\title{Print the output from a \code{coxivtsrif} model fit.}
\usage{
\method{print}{coxivtsrif}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of the class "\code{coxivtsrif}", usually created by \code{coxiv_tsrif()}.}

\item{digits}{the number of significant digits to display for coefficients and hazard ratios.}

\item{...}{additional arguments (currently unused).}
}
\description{
Provide succinct and organized output for a \code{coxivtsrif} model, such as \code{coxiv_tsrif()}.
}
\examples{
print(coxiv_tsrif(Surv(time, censor) ~ x, data = dat))

}
