% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.seqem.R
\name{print.seqem}
\alias{print.seqem}
\title{Print method for seqem objects}
\usage{
\method{print}{seqem}(x, ...)
}
\arguments{
\item{x}{An object of class \code{seqem} returned by \code{\link[=seqem]{seqem()}}.}

\item{...}{Ignored.}
}
\description{
Print method for seqem objects
}
