% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.surviv.R
\name{print.summary.surviv}
\alias{print.summary.surviv}
\title{Print the output from a \code{surviv} model fit.}
\usage{
\method{print}{summary.surviv}(object, digits = max(1L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{a \code{summary.surviv} object.}
}
\description{
This method will print a better-looking output for summary of a \code{surviv} model object.
}
\examples{
print(summary(CoxIVeq(Surv(time, censor) ~ x, data = dat)))

}
