% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat_omom.R
\name{simdat_omom}
\alias{simdat_omom}
\title{Simulate toy data for MOM IV estimation of Cox model}
\usage{
simdat_omom(n, effect)
}
\arguments{
\item{n}{sample size specified by user.}

\item{effect}{user-specified causal effect in terms of log(HR) or beta.}
}
\value{
a data frame contains survival time, censor indicator, treament, IV, and covariates/
}
\description{
Simulate survival data with treatment ("trt"), three covariates ("x1", "x2", "x3), and an instrumental variable ("IV").
The outcome is survival time ("Tiempo") and censorship ("censor"). Only one covariate is confounder.
}
\examples{
simdat_omom(0.5)
}
