% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat_tvcoef.R
\name{simdat_tvcoef}
\alias{simdat_tvcoef}
\title{Generate a data set for test purposes or to conduct a simulation study}
\usage{
simdat_tvcoef(
  n,
  p,
  q,
  efftrtsurv,
  effcovsurv,
  effusurv,
  effcovtrt,
  effivtrt,
  effvtrt
)
}
\arguments{
\item{n}{Number of data values}

\item{p}{Number of exogeneous covariates}

\item{q}{Number of IVs}

\item{efftrtsurv}{Treatment effect on survival time}

\item{effcovsurv}{Observed confounders effect on survival time}

\item{effusurv}{Unmeasured confounder effect on survival time}

\item{effcovtrt}{Observed covariates effect on treatment selection}

\item{effivtrt}{Instrumental variables effect on treatment selection}

\item{effvtrt}{Unmeasured confounder effect on treatment selection}
}
\value{
Generated data and full info as list
}
\description{
Generate a data set for test purposes or to conduct a simulation study
}
