% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.surviv.R
\name{summary.surviv}
\alias{summary.surviv}
\title{Summarizing Cox Regression using Instrumental Variable}
\usage{
\method{summary}{surviv}(object, ...)
}
\arguments{
\item{object}{an object of the class "\code{surviv}", usually a call from "\code{coxiv_mom}".}
}
\description{
Provide summary for a \code{surviv} object.
}
\examples{
summary(coxiv_mom(Surv(time, censor) ~ x, data = dat))

}
